+++ src/Context.js
@@ -3,12 +3,14 @@ import Environment from './Environment.js';
 import ExtensionsMapper from './ExtensionsMapper.js';
 import {getUniqueId} from './shared.js';
 
+const kOwner = Symbol.for('owner');
+
 export default function Context(definitionContext, environment) {
   environment = environment ? environment.clone() : new Environment();
   return new ContextInstance(definitionContext, environment);
 }
 
-function ContextInstance(definitionContext, environment) {
+function ContextInstance(definitionContext, environment, owner) {
   const {id = 'Def', name, type = 'context'} = definitionContext;
   const sid = getUniqueId(id);
   this.id = id;
@@ -29,7 +31,14 @@ function ContextInstance(definitionContext, environment) {
     sequenceFlowRefs: {},
     sequenceFlows: [],
   };
+  this[kOwner] = owner;
 }
+
+Object.defineProperty(ContextInstance.prototype, 'owner', {
+  get() {
+    return this[kOwner];
+  },
+});
 
 ContextInstance.prototype.getActivityById = function getActivityById(activityId) {
   const activityInstance = this.refs.activityRefs[activityId];
@@ -106,8 +115,8 @@ ContextInstance.prototype.upsertAssociation = function upsertAssociation(associa
   return instance;
 };
 
-ContextInstance.prototype.clone = function clone(newEnvironment) {
-  return new ContextInstance(this.definitionContext, newEnvironment || this.environment);
+ContextInstance.prototype.clone = function clone(newEnvironment, newOwner) {
+  return new ContextInstance(this.definitionContext, newEnvironment || this.environment, newOwner);
 };
 
 ContextInstance.prototype.getProcessById = function getProcessById(processId) {
@@ -120,6 +129,8 @@ ContextInstance.prototype.getProcessById = function getProcessById(processId) {
 
   const bpContext = this.clone(this.environment.clone());
   bp = refs[processId] = new processDefinition.Behaviour(processDefinition, bpContext);
+  bpContext[kOwner] = bp;
+
   this.refs.processes.push(bp);
 
   return bp;
@@ -128,7 +139,11 @@ ContextInstance.prototype.getProcessById = function getProcessById(processId) {
 ContextInstance.prototype.getNewProcessById = function getNewProcessById(processId) {
   if (!this.getProcessById(processId)) return null;
   const bpDef = this.definitionContext.getProcessById(processId);
-  const bp = new bpDef.Behaviour(bpDef, this.clone(this.environment.clone()));
+
+  const bpContext = this.clone(this.environment.clone());
+  const bp = new bpDef.Behaviour(bpDef, bpContext);
+  bpContext[kOwner] = bp;
+
   return bp;
 };
 
@@ -200,4 +215,12 @@ ContextInstance.prototype.loadExtensions = function loadExtensions(activity) {
   if (io.hasIo) extensions.extensions.push(io);
   if (!extensions.extensions.length) return;
   return extensions;
+};
+
+ContextInstance.prototype.getActivityParentById = function getActivityParentById(activityId) {
+  const owner = this[kOwner];
+  if (owner) return owner;
+  const activity = this.getActivityById(activityId);
+  const parentId = activity.parent.id;
+  return this.getProcessById(parentId) || this.getActivityById(parentId);
 };