+++ src/activity/Activity.js
@@ -87,6 +87,7 @@ function Activity(Behaviour, activityDef, context) {
     isTransaction: activityDef.isTransaction,
     isParallelJoin,
     isThrowing: activityDef.isThrowing,
+    lane: activityDef.lane && activityDef.lane.id,
   };
   this[kExec] = {};
 
@@ -235,10 +236,27 @@ Object.defineProperty(Activity.prototype, 'attachedTo', {
   },
 });
 
+Object.defineProperty(Activity.prototype, 'lane', {
+  enumerable: true,
+  get() {
+    const laneId = this[kFlags].lane;
+    if (!laneId) return undefined;
+    const parent = this.parentElement;
+    return parent.getLaneById && parent.getLaneById(laneId);
+  },
+});
+
 Object.defineProperty(Activity.prototype, 'eventDefinitions', {
   enumerable: true,
   get() {
     return this[kEventDefinitions];
+  },
+});
+
+Object.defineProperty(Activity.prototype, 'parentElement', {
+  enumerable: true,
+  get() {
+    return this.context.getActivityParentById(this.id);
   },
 });