+++ src/process/Process.js
@@ -10,6 +10,7 @@ const kCounters = Symbol.for('counters');
 const kExec = Symbol.for('execution');
 const kExecuteMessage = Symbol.for('executeMessage');
 const kExtensions = Symbol.for('extensions');
+const kLanes = Symbol.for('lanes');
 const kMessageHandlers = Symbol.for('messageHandlers');
 const kStateMessage = Symbol.for('stateMessage');
 const kStatus = Symbol.for('status');
@@ -53,6 +54,9 @@ export function Process(processDef, context) {
 
   this.logger = environment.Logger(type.toLowerCase());
 
+  if (behaviour.lanes) {
+    this[kLanes] = behaviour.lanes.map((lane) => new lane.Behaviour(this, lane));
+  }
   this[kExtensions] = context.loadExtensions(this);
 }
 
@@ -60,6 +64,14 @@ Object.defineProperty(Process.prototype, 'counters', {
   enumerable: true,
   get() {
     return {...this[kCounters]};
+  },
+});
+
+Object.defineProperty(Process.prototype, 'lanes', {
+  enumerable: true,
+  get() {
+    const lanes = this[kLanes];
+    return lanes && lanes.slice();
   },
 });
 
@@ -397,6 +409,12 @@ Process.prototype.getSequenceFlows = function getSequenceFlows() {
   const execution = this.execution;
   if (execution) return execution.getSequenceFlows();
   return this.context.getSequenceFlows();
+};
+
+Process.prototype.getLaneById = function getLaneById(laneId) {
+  const lanes = this[kLanes];
+  if (!lanes) return;
+  return lanes.find((lane) => lane.id === laneId);
 };
 
 Process.prototype.getPostponed = function getPostponed(...args) {